Fix the bug :shipit:
