+++ packages/next/shared/lib/amp-mode.ts
@@ -1,0 +1,7 @@
+export function isInAmpMode({
+  ampFirst = false,
+  hybrid = false,
+  hasQuery = false,
+} = {}): boolean {
+  return ampFirst || (hybrid && hasQuery)
+}
