+++ packages/next/shared/lib/amp.ts
@@ -1,13 +1,6 @@
 import React from 'react'
 import { AmpStateContext } from './amp-context'
-
-export function isInAmpMode({
-  ampFirst = false,
-  hybrid = false,
-  hasQuery = false,
-} = {}): boolean {
-  return ampFirst || (hybrid && hasQuery)
-}
+import { isInAmpMode } from './amp-mode'
 
 export function useAmp(): boolean {
   // Don't assign the context value to a variable to save bytes
