+++ packages/next/shared/lib/head.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react'
 import Effect from './side-effect'
 import { AmpStateContext } from './amp-context'
 import { HeadManagerContext } from './head-manager-context'
-import { isInAmpMode } from './amp'
+import { isInAmpMode } from './amp-mode'
 import { warnOnce } from './utils'
 
 type WithInAmpMode = {
